task_2.1
