+++ 2015/NatashaMartinchik/2st_task/Gemfile
@@ -3,3 +3,6 @@ source "https://rubygems.org"
 
 
 gem 'colored'
+gem 'colored'
+gem 'json'
+gem 'pry'
