+++ 2015/NatashaMartinchik/2st_task/Gemfile~
@@ -1,0 +1,5 @@
+
+source "https://rubygems.org"
+
+
+gem 'colored'
