+++ 2015/NatashaMartinchik/2st_task/lib/Visualizer.rb
@@ -1,0 +1,9 @@
+class Visualizer
+  def initialize(result)
+    @result = result
+  end
+
+  def visual
+    @result.each { |v| puts (@result.include?(v) ? v.red : v) }
+  end
+end
